fixes latest release that removed the store from the props passed from
