+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "redux-zero",
-  "version": "4.1.0",
+  "version": "4.1.1",
   "description": "",
   "main": "dist/redux-zero.js",
   "typings": "dist/index.d.ts",
